Bugs and reorg
AI was moved again to the System.updateall() function, and seems to be fixed.  Destroy mission generation was removed from System.randomplanets() function and moved to System.enemypopulate(), so it now can't happen without a system full of bots to pick from.  I think it can still theoretically call for a hit on the player, but I haven't seen it happen and kind of want to.