Outpost and AI tweaks
Guaranteed 2 weapons to buy in every outpost (although 1/10 chance they are the same weapon listed twice), cleaner code selects from an array of appropriate Shopitems rather than creating new ones.  AI was tweaked to prevent shooting at player on the far side of it's parent planet, but allow shooting at the player on the near side of it's parent planet.